% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_slope_plane_fit.R
\name{align_slope_plane_fit}
\alias{align_slope_plane_fit}
\title{3D plane fit for a set of points}
\usage{
align_slope_plane_fit(terrain_data = "*.xyz", full_data = "*.xyz",
  plotting = F)
}
\arguments{
\item{terrain_data}{A file containing the (terrain) points into which the plane will be fitted}

\item{full_data}{A file containing the points that will be aligned/rotation based on the plane surface normal}
}
\value{
Slope in degree of the aligned data.
}
\description{
Fits a 3D plane into a set of 3D points (x,y,z)
}
\examples{
align_slope_plane_fit(terrain_data="terrain.xyz", full_data="full.xyz", plotting=F)
}
\author{
Matthias Kunz, last updated: 18.12.2019
}
