% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashape2vox.R
\name{ashape2vox}
\alias{ashape2vox}
\title{Voxelization of an 3D alphashape object}
\usage{
ashape2vox(input_ashape3d = "", voxel_size = 0.2)
}
\arguments{
\item{input_ashape3d}{Tree crowns ashape3d object based on alphashape3d R package (see get_3d_alpha_shape)}

\item{voxel_size}{Size of the voxels (in meter) of the underlying voxel grid. Default 0.2}
}
\value{
Simple voxelgrid point cloud (X, Y, Z) of all voxels inside the given 3D alphashape.
}
\description{
Voxelization of an 3D alphashape crown model into a voxelgrid model
}
\examples{

}
\author{
Carsten Hess, last updated: 10.02.2020
}
