% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbh_from_qsm.R
\name{cbh_from_qsm}
\alias{cbh_from_qsm}
\title{Crown base height (CBH) extraction from a quantitative structure model (QSM).}
\usage{
cbh_from_qsm(cyl_data = "cyl_data_*.txt", min_diameter_m = 0.07,
  max_angle_deg = 60, min_branch_length_m = 2)
}
\arguments{
\item{cyl_data}{A file containing a QSM, e.g. cyl_data_*.txt}

\item{max_angle_deg}{Maximum allowed branch angle (in degree) from vertical tree axis. Default 60.0}

\item{min_branch_length_m}{Minimum length of a first order branch considered to be first living trunk. Default 2.0}

\item{min_diameter}{Minimum diameter (in meter) of a first order branch considered to be first living trunk. Default 0.07}
}
\value{
CBH in meter. Note that CBH computation sets lowest tree point in z axis to 0.
}
\description{
Extracts the CBH from a QSM file. QSM file (cyl_data*.txt) must come from TreeQSM 2.30.
}
\examples{
cbh_from_qsm("cyl_data_*.txt", 0.07, 60.0, 2.0)
}
\references{
\url{https://github.com/InverseTampere/TreeQSM}
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
