% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qsm_multistem.R
\name{check_qsm_multistem}
\alias{check_qsm_multistem}
\title{Correction of QSM for multistem trees.}
\usage{
check_qsm_multistem(qsm = "cyl_data*.txt", bifurcation_threshold = 1.5,
  percentage = 40, diameter_treshold = 0.035, min_tree_height = 2,
  plot_orders = F, plot_parts = F, replace_qsm = F)
}
\arguments{
\item{qsm}{A file containing a QSM, e.g. cyl_data_*.txt}

\item{bifurcation_threshold}{Height threshold (in meter) under which bifurcation of an additional trunk is allowed. Default 1.5}

\item{percentage}{Percentage difference in radius between potential new trunk and main trunk. Default 40.0}

\item{diameter_treshold}{Minimum diameter (in meter) threhold for potential additional trunks. Default 0.035}

\item{min_tree_height}{Minimum tree height for searching for additional trunks. Default 2.0}

\item{plot_orders}{Plot branching orders original and checked QSM. Default to FALSE}

\item{plot_parts}{QSM for each detected stem (column tree_part). Default to FALSE}

\item{replace_qsm}{Should input QSM file be replaced. If FALSE _multistem.txt is added to file name. Default to FALSE}
}
\value{
The path to the corrected QSM file.
}
\description{
Correction of QSM for multistem trees.
}
\examples{
check_qsm_multistem("cyl_data*.txt", 1.5, 40.0, 0.035, 2.0, FALSE, FALSE, FALSE)
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
