% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Radian to degree conversion.}
\usage{
rad2deg(angle = 0)
}
\arguments{
\item{angle}{A positive angle in radian (0-1)}
}
\value{
The angle in degree (0-360)
}
\description{
Radian to degree conversion.
}
\section{Functions}{
\itemize{
\item \code{rad2deg}: an angle in radian (0-1) to degree (0-360)
}}

\examples{
rad2deg(0.7853982) # 45.0
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
