% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_complementarity.R
\name{crown_complementarity}
\alias{crown_complementarity}
\title{Computation of crown complementarity based on slice volumes.}
\usage{
crown_complementarity(tree_i = NA, tree_j = NA, strata_height = 0.3,
  vox = FALSE, voxel_size = 0.03, plotting = FALSE)
}
\arguments{
\item{tree_i}{A file or data.frame containing a tree point cloud with x y z}

\item{tree_j}{A file or data.frame containing a tree point cloud with x y z}

\item{strata_height}{Height of strata (in meter) for which the complementarity is computed. Default 0.3}

\item{vox}{Should the input point cloud be reduced to voxel grid for faster computation. Default to TRUE.}

\item{voxel_size}{Size (in meters) of the voxels when vox==TRUE. Default 0.03}

\item{plotting}{Should strata be plotted. Default to FALSE}
}
\value{
The crown complementarity value (CC_ij)
}
\description{
Correction of QSM for multistem trees.
}
\examples{
crown_complementarity(tree_i, tree_j, strata_height=0.3, plotting=FALSE)
}
\references{
Williams et al., NEE, 2017, DOI: 10.1038/s41559-016-0063
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
