% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder3d.R
\name{cylinder3d}
\alias{cylinder3d}
\title{Visualise a 3D cylinder}
\usage{
cylinder3d(p_start = c(0, 0, 0), orient = c(1, 1, 1), radius = 0.25,
  length = 1, order = 0, n_seg = 12, axis = F, cylinders = T,
  circles = F, transparent = T)
}
\arguments{
\item{p_start}{A vector with the start point of the cylinder in 3D (x,y,z)}

\item{orient}{Directional vector of the cylinder axis (x1,y1,z1)}

\item{radius}{Radius of the cylinder (in meter)}

\item{length}{Length of the cylinder (in meter)}

\item{order}{Branch order level of the cylinder. This is optional for coloring. Default 0}

\item{n_seg}{Number of segments that draw the cylinder. Default 12}

\item{axis}{Display the cylinder axis. Default to FALSE.}

\item{cylinders}{Display the cylinder. Defaut to TRUE.}

\item{circles}{Display circles at the end of the cylinder. Default to FALSE.}

\item{transparent}{Makes the cylinder transparent. Default to TRUE with transparency of 0.3}
}
\value{
Does not return anything at the moment.
}
\description{
Visualise a 3D cylinder
}
\examples{
cylinder3d(p_start = c(0,0,0), orient = c(1,1,1), radius = 0.25, length = 1)
}
\author{
Matthias Kunz, last updated: 29.04.2019
}
