% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Degree to radian conversion.}
\usage{
deg2rad(angle = 0)
}
\arguments{
\item{angle}{A positive angle in degree (0-360)}
}
\value{
The angle in radian
}
\description{
Converts an angle in degree (0-360) to radian (0-1)
}
\examples{
deg2rad(45.0) # 0.7853982
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
