% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm2vox.R
\name{dgm2vox}
\alias{dgm2vox}
\title{Voxelization of a set of ground points (DGM).}
\usage{
dgm2vox(input_cloud = "*.xyz", voxel_size = 0.1, raster_size = 0.2)
}
\arguments{
\item{voxel_size}{Size of the voxel edge length (in meter) of the voxelgrid ground points (DGM). Default 0.1}

\item{raster_size}{Size of the rastergrid (in meter) of preprocessed ground points (DGM) to further eliminate possible noise points at excessive ground level.}

\item{input_dgm}{A file or data.frame containing a set ground points with x y z. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}
}
\value{
Voxelgrid of DGM with closed/interploated gaps due to occlusions.
}
\description{
Converts cloud of ground points (x y z) to a voxelgrid model with defined voxel size. Gaps/missing points in the voxelgrid were interpolated.
}
\examples{

}
\author{
Carsten Hess, last updated: 10.02.2020
}
