% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_percent.R
\name{diff_percent}
\alias{diff_percent}
\title{Compute absolute difference in percent between two values}
\usage{
diff_percent(x1, x2)
}
\arguments{
\item{x1}{A value}

\item{x2}{Another value}
}
\value{
Difference between x1 and x2 in absolute (always positive) percent
}
\description{
Compute absolute difference in percent between two values
}
\examples{
diff_percent(10,12) # 18.18\%
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
