% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_cbh.R
\name{display_cbh}
\alias{display_cbh}
\title{Display tree crown base height.}
\usage{
display_cbh(input_cloud = "*.xyz", cbh = 0, vox = TRUE,
  voxel_size = 0.03)
}
\arguments{
\item{input_cloud}{A 3D point cloud file, e.g. "tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{cbh}{Crown base height in meter. Default 0.0}

\item{vox}{Should the input point cloud be displayed as voxel grid for faster display. Default to TRUE.}

\item{voxel_size}{Size (in meters) of the voxels when vox==TRUE. Default 0.03}
}
\value{
NULL. Function currently does not return anything.
}
\description{
Display tree crown base height.
}
\examples{
display_cbh("tree1.xyz", 3.4)
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
