% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_qsm.R
\name{display_qsm}
\alias{display_qsm}
\title{Visualize 3D QSM model}
\usage{
display_qsm(qsm = "cyl_data*.txt", n_seg = 12, axis = FALSE,
  cylinders = TRUE, circles = FALSE, transparent = TRUE)
}
\arguments{
\item{qsm}{qsm A file containing a QSM, e.g. cyl_data_*.txt}

\item{n_seg}{Number of segments that draw the cylinder. Default 12}

\item{axis}{Display the cylinder axis. Default to FALSE.}

\item{cylinders}{Display the cylinder. Defaut to TRUE.}

\item{circles}{Display circles at the end of the cylinder. Default to FALSE.}

\item{transparent}{Makes the cylinder transparent. Default to TRUE with transparency of 0.3}
}
\value{
Does not return anything at the moment.
}
\description{
Visualize 3D QSM model from TreeQSM 2.30
}
\examples{
display_qsm("cyl_data*.txt")
}
\references{
\url{https://github.com/InverseTampere/TreeQSM}
}
\author{
Matthias Kunz, last updated: 29.04.2019
}
