% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENL.R
\name{ENL}
\alias{ENL}
\title{Computes the effective number of layers (ENL) after Ehbrecht et al. 2016}
\usage{
ENL(input_cloud = df, voxel_size = 0.05, slice_thickness = 0.5)
}
\arguments{
\item{input_cloud}{A 3D point cloud file, e.g. "tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{voxel_size}{Voxel size (in meter) for ENL computation. Default 0.1 m.}

\item{slice_thickness}{Thickness (in meter) of vertical layers for ENL computation. Default 0.5 m.}
}
\value{
ENL value
}
\description{
Computes the effective number of layers (ENL) after Ehbrecht et al. 2017
}
\examples{
ENL(input_cloud=df, voxel_size=0.05, slice_thickness=0.5)
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
