% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENL_height_normalized.R
\name{ENL_height_normalized}
\alias{ENL_height_normalized}
\title{Computes the effective number of layers (ENL) after Ehbrecht et al. 2016}
\usage{
ENL_height_normalized(input_cloud = ".xyz", voxel_size = 0.1,
  slice_thickness = 0.5)
}
\arguments{
\item{input_cloud}{A 3D point cloud file as full path, e.g. "C:/tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{voxel_size}{Voxel size (in meter) for ENL computation. Default 0.1 m.}

\item{slice_thickness}{Thickness (in meter) of vertical layers for ENL computation. Default 0.5 m.}
}
\value{
ENL value
}
\description{
Computes the effective number of layers (ENL) after Ehbrecht et al. 2017
}
\examples{
ENL(input_cloud="*.xyz", voxel_size=0.05, slice_thickness=0.5)
}
\author{
Matthias Kunz, last updated: 28.02.2020
}
