% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_distribution.R
\name{fit_beta_distribution}
\alias{fit_beta_distribution}
\title{Fit beta distribution.}
\usage{
fit_beta_distribution(profile = c(1, 1.1, 1))
}
\arguments{
\item{profile}{A 3D point cloud file, e.g. "tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}
}
\value{
Beta distribution
}
\description{
Print/Display tree crown projection area.
}
\examples{
fit_beta("tree1.xyz")
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
