% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_2d_alpha_shape.R
\name{get_2d_alpha_shape}
\alias{get_2d_alpha_shape}
\title{Compute area, perimeter and centroid of an 2d alpha-shape for a set of points.}
\usage{
get_2d_alpha_shape(
  input_cloud = data.frame(x = runif(20), y = runif(20)),
  alpha = 1,
  plot = FALSE
)
}
\arguments{
\item{input_cloud}{A file or data.frame containing x and y points. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{alpha}{Alpha value for alpha-shape computation. Default 0.5}

\item{plot}{Should the alpha shape be plotted. Default to FALSE.}
}
\value{
A data.frame containing: area, perimeter, cpa_cen_x, cpa_cen_y.
}
\description{
Compute area, perimeter and centroid of an 2d alpha-shape.\cr
    Alpha-shape is checked for connectivity and circular graphs.\cr
    If problems occur alpha-value is increased by 0.2. To avoid collinearity a small random jitter is added to the points before alpha-shape computation.
}
\examples{
get_2d_alpha_shape(data.frame(x=runif(20),y=runif(20)), alpha=1.0)
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
