% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_2d_chull.R
\name{get_2d_chull}
\alias{get_2d_chull}
\title{Compute area, perimeter and centroid of an 2d convex hull for a set of points.}
\usage{
get_2d_chull(input_cloud = data.frame(x = runif(20), y = runif(20)),
  plot = FALSE)
}
\arguments{
\item{input_cloud}{A file or data.frame containing x and y points. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{plot}{Should the convex hull be plotted. Default to FALSE.}
}
\value{
A data.frame containing: area, perimeter, cpa_cen_x, cpa_cen_y.
}
\description{
Compute area, perimeter and centroid of an 2d convex hull.
}
\examples{
get_2d_chull(data.frame(x=runif(20),y=runif(20)))
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
