% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3d_alpha_shape.R
\name{get_3d_alpha_shape}
\alias{get_3d_alpha_shape}
\title{Computes the 3d alpha shape, volume, surface area, and centroid of a 3D set of points}
\usage{
get_3d_alpha_shape(input_cloud = data.frame(x = runif(20), y = runif(20),
  z = runif(20)), alpha = 1, plot = FALSE)
}
\arguments{
\item{input_cloud}{A file or data.frame containing a tree point cloud with x y z. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{alpha}{Alpha value for alpha-shape computation. Default 0.5}

\item{plot}{Should the 3d alpha shape be plotted. Default to FALSE.}
}
\value{
A data.frame containing: volume, surface area, vol_cen_x, vol_cen_y, vol_cen_z, alpha value and the alpha-shape object.
}
\description{
computes the 3d alpha shape, volume, surface area, and centroid of a 3D set of points\cr
    To avoid collinearity a small random jitter is added to the points before alpha-shape computation.
}
\examples{
get_3d_alpha_shape(data.frame(x=runif(20),y=runif(20), z=runif(20)), alpha=1.0)
}
\author{
Matthias Kunz, last updated: 29.04.2019
}
