% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circle_fit.R
\name{get_circle_fit}
\alias{get_circle_fit}
\title{Computes the 3d alpha shape, volume, surface area, and centroid of a 3D set of points}
\usage{
get_circle_fit(
  input_cloud = data.frame(x = runif(20), y = runif(20), z = runif(20)),
  plot_fit = FALSE
)
}
\arguments{
\item{input_cloud}{A file or data.frame containing a tree point cloud with x y z. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{plot_fit}{Should the circle fit be plotted. Default to FALSE.}
}
\value{
A .txt files containing: CPA, CH, Height, GD, DBH, etc.
}
\description{
computes the 2d alpha shape, 2d convex hull a 3D set of points as well as ground diameter and diameter at breast height\cr
    To avoid collinearity a small random jitter is added to the points before alpha-shape computation.
}
\examples{
get_circle_fit(data.frame(x=runif(20),y=runif(20), z=runif(20)), plot=FALSE)
}
\author{
Matthias Kunz, last updated: 26.11.2019
}
