% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shadow_voxel.R
\name{get_shadow_voxel}
\alias{get_shadow_voxel}
\title{Simple identification of shadow voxel inside the stem segment of the tree voxelmodel.}
\usage{
get_shadow_voxel(input_vox = NA, voxel_size = 0.2)
}
\arguments{
\item{input_vox}{Voxelmodel of the trees stem segment (X,Y,Z)}
}
\value{
Voxelmodel with added and marked (TRUE, FALSE) shadow voxel.
}
\description{
If the voxel size is much smaller than DBH a simple voxelization will not reconsider the space inside the stem.
}
\examples{

}
\author{
Carsten Hess, last updated: 14.02.2020
}
