% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanFrac.R
\name{MeanFrac}
\alias{MeanFrac}
\title{Computation of mean fractal dimensions for a point cloud with definded cross section based on angular resolution}
\usage{
MeanFrac(input_cloud = df, angular_resolution = 360/2560,
  plotting = F)
}
\arguments{
\item{angular_resolution}{Angular resolution of the scan, i.e. 360/number of points. Default 360/2560.}

\item{plotting}{Logical. Should the MeanFrac polygons be plotted. Note: This may takes a long time. Default FALSE.}

\item{input_df}{Data frame that holds the point cloud}
}
\value{
List with MeanFrac and fractal dimension for each cross section
}
\description{
Computation of mean fractal dimensions for a point cloud with definded cross section based on angular resolution
}
\examples{
MeanFrac(input_cloud=df, angular_resolution=360/2560, plotting=F)
}
\author{
Matthias Kunz, last updated: 19.12.2019
}
