% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CPA.R
\name{print_CPA}
\alias{print_CPA}
\title{Display CPA.}
\usage{
print_CPA(input_cloud = "*.xyz")
}
\arguments{
\item{input_cloud}{A 3D point cloud file, e.g. "tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}
}
\value{
Return a PNG file
}
\description{
Print/Display tree crown projection area.
}
\examples{
print_CPA("tree1.xyz")
}
\author{
Matthias Kunz, last updated: 27.04.2019
}
