% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tree_analysis.R
\name{single_tree_analysis}
\alias{single_tree_analysis}
\title{Computes the 3d alpha shape, volume, surface area, and centroid of a 3D set of points}
\usage{
single_tree_analysis(input_cloud = data.frame(x = runif(20), y =
  runif(20), z = runif(20)), vox = FALSE, voxel_size = 0.03,
  alpha_value = 1, plot = FALSE, plot_hulls = FALSE,
  overwrite_output = TRUE)
}
\arguments{
\item{input_cloud}{A file or data.frame containing a tree point cloud with x y z. If file, it assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{vox}{Should the input point cloud be reduced to voxel grid for faster computation. Default to TRUE.}

\item{voxel_size}{Size (in meters) of the voxels when vox==TRUE. Default 0.03}

\item{alpha_value}{Alpha value. Default 1.0}

\item{plot}{Should the 2d alpha shape be plotted. Default to FALSE.}

\item{plot_hulls}{Should the CPA and slice hull be plotted. Default to FALSE.}

\item{overwrite_output}{Should outout files be overwritten?. Logical. Default TRUE.}
}
\value{
A .txt files containing: CPA, CH, Height, GD, DBH, etc..
}
\description{
computes the 2d alpha shape, 2d convex hull a 3D set of points as well as ground diameter and diameter at breast height\cr
    To avoid collinearity a small random jitter is added to the points before alpha-shape computation.
}
\examples{
get_3d_alpha_shape(data.frame(x=runif(20),y=runif(20), z=runif(20)), alpha=1.0)
}
\author{
Matthias Kunz, last updated: 26.11.2019
}
