% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre3d-package.R
\docType{package}
\name{tre3d-package}
\alias{tre3d-package}
\title{tre3d: An R-package for computating tree metrics from 3D point clouds.}
\description{
The tre3d package provides some cool functions:
cbh_from_qsm, crown_metrics and pointcloud2voxelgrid.
}
\section{tre3d functions}{

cbh_from_qsm\cr
crown_metrics\cr
pointcloud2voxelgrid
}

\author{
Matthias Kunz, matthias.kunz[at]tu-dresden.de, last updated: 13.12.2019
}
