% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre3d-package.R
\docType{package}
\name{tre3d}
\alias{tre3d}
\alias{tre3d-package}
\title{tre3d: A package for computating tree metrics from point clouds.}
\description{
The tre3d package provides some cool functions:
cbh_from_qsm, crown_metrics and pointcloud2voxelgrid.
}
\section{tre3d functions}{

cbh_from_qsm\cr
crown_metrics\cr
pointcloud2voxelgrid
}

\author{
Matthias Kunz, last updated: 28.04.2019
}
