% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre3d-version.R
\name{tre3d_version}
\alias{tre3d_version}
\title{tre3d version information}
\usage{
tre3d_version()
}
\description{
Gives information of the tre3d version and changelog
}
\examples{
tre3d_version()
}
\author{
Matthias Kunz, last updated: 29.02.2020
}
