% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree1.R
\docType{data}
\name{tree1}
\alias{tree1}
\title{tree1}
\format{A data frame with n rows and 3 variables:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{z}{z coordinate}
}}
\usage{
data(grav)
}
\description{
A point cloud of a single pinguin
}
\keyword{datasets}
