% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_end.R
\name{vector_end}
\alias{vector_end}
\title{Compute end point of given point in 3d with direction and length}
\usage{
vector_end(p_start = c(0, 0, 0), orient = c(1, 1, 1), length = 1)
}
\arguments{
\item{p_start}{A vector with the start point of the vector, e.g. a cylinder, in 3D (x,y,z)}

\item{orient}{Directional vector of the vector, e.g. a cylinder, axis (x1,y1,z1)}

\item{length}{Length of the vector, e.g. cylinder length (in meter)}
}
\value{
The end point (p_end) of the vector in 3D (x,y,z)
}
\description{
Compute end point of given point in 3d with direction and length
}
\examples{
vector_end(p_start = c(0,0,0), orient = c(1,1,1), length = 1)
}
\author{
Matthias Kunz, last updated: 29.04.2019
}
