% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_distribution.R
\name{vertical_distribution}
\alias{vertical_distribution}
\title{Vertical_distribution.}
\usage{
vertical_distribution(input_cloud = data.frame(x = runif(20), y =
  runif(20), z = runif(20)), strata_height = 0.3, vox = TRUE,
  voxel_size = 0.03)
}
\arguments{
\item{input_cloud}{A 3D point cloud file, e.g. "tree1.xyz".
Assumes first three columns to be x y z without header. Futher columns are ignored.}

\item{strata_height}{Thickness of vertical slices in meter. Default 0.3}

\item{vox}{Should the input point cloud be reduced to voxel grid for faster computation. Default to TRUE.}

\item{voxel_size}{Size (in meters) of the voxels when vox==TRUE. Default 0.03}
}
\value{
A data.frame with tree parameters
}
\description{
Computation of vertical (wood) distribution
}
\examples{
vertical_distribution("tree1.xyz", 0.25)
}
\author{
Matthias Kunz, last updated: 30.01.2020
}
