% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelgrid_plot_model.R
\name{voxelgrid_plot_model}
\alias{voxelgrid_plot_model}
\title{Create a voxelgrid datatable on plot level to further analyse of the spatial canopy occupation.}
\usage{
voxelgrid_plot_model(
  plot_id = "Dummy",
  path_voxelgrid_folder = NA,
  voxel_size = 0.2,
  clip_area = TRUE
)
}
\arguments{
\item{plot_id}{Name/ID of the plot/forest stand}

\item{path_voxelgrid_folder}{Plot based path to the folder which contains all relevant tree and DGM voxelgrid models.}

\item{clip_area}{Clip the underlying dgm to the spatial extend of the projected tree voxelgrid models}
}
\value{
Saved datagrid table which contains all plot based voxelgrid models and data.table object with total voxel volumes for each type of occupation.
}
\description{
Join the voxelgrid models of the DGM and all trees as one big voxelgrid datatable.
}
\examples{

}
\author{
Carsten Hess, last updated: 20.02.2020
}
