% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelgrid_tree_model.R
\name{voxelgrid_tree_model}
\alias{voxelgrid_tree_model}
\title{Create a voxelgrid tree model including the voxel attributs tree height, segment (crown, stem) and type of occupation (pcd, ashape3d)}
\usage{
voxelgrid_tree_model(tree_id = "Dummy", pcd_vox = NA, as3d_vox = NA,
  alpha_value = 0.5, voxel_size = 0.2)
}
\arguments{
\item{pcd_vox}{Voxelmodel of the tree point cloud (Size of the voxels (see pcd2vox)}

\item{as3d_vox}{Voxelmodel of the tree crown 3D alphashape (see get_3d_alpha_shape)}

\item{alpha_value}{Documented modelparameter of the pre-processed tree crown 3D alphashape (ashaped3d).}

\item{voxel_size}{Documented modelparameter of the pre-prcessed voxelmodels (pcd, ashape).}
}
\value{
List object of tree with underlying modelparameter and Voxelgrid tree model (X,Y,Z,type)
}
\description{
Merge voxelmodels of tree point cloud and crown 3D alphashape.
}
\examples{

}
\author{
Carsten Hess, last updated: 10.02.2020
}
