% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz2cube3d.R
\name{xyz2cube3d}
\alias{xyz2cube3d}
\title{Visualise 3d point as voxel}
\usage{
xyz2cube3d(x = 0, y = 0, z = 0, bordered = TRUE, filled = TRUE,
  lwd = 2, scale = 1, fillcol = "gray", bordercol = "black",
  alpha = 0.5, ...)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{bordered}{Logical, Draws a border}

\item{filled}{Logical, Fills the cube with a color}

\item{lwd}{Line width of the border. Default 2}

\item{scale}{Size of the cube, e.g. 0.1 when 10 cm voxels are used. Default 1 (meter).}

\item{fillcol}{## Color of the cube. Default gray.}

\item{bordercol}{Color of the border}

\item{alpha}{Transparancy of the cube}
}
\value{
A cube object that can be plotted with rgl
}
\description{
Visualise 3d point as voxel. Usage with rgl: open3d(); cube3d(1,1,1,...)
}
\examples{
xyz2cube3d(data.frame(x=runif(20),y=runif(20), z=runif(20)), alpha=1.0)
}
\author{
Matthias Kunz, last updated: 18.02.2020
}
