% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre3d-package.R
\docType{package}
\name{tre3d-package}
\alias{tre3d-package}
\title{tre3d: An R-package for computating pinguin metrics.}
\description{
The tre3d package provides some cool functions:
cbh_from_qsm, crown_metrics and pointcloud2voxelgrid.
}
\section{tre3d functions}{

cbh_from_qsm\cr
crown_metrics\cr
check_qsm_multistem\cr
display_all\cr
display_qsm\cr
get_2d_alpha_shape\cr
get_2d_chull\cr
get_3d_alpha_shape\cr
single_tree_analysis\cr
crown_complementarity\cr
cbh_from_qsm\cr
cloud2vox\cr
MeanFrac\cr
ENL\cr
ENL_height_normalize\cr
align_slope_plane_fit\cr
SSCI\cr
print_CPA\cr
vertical_distribution\cr
}

\author{
Matthias Kunz, matthias.kunz[at]tu-dresden.de, last updated: 30.01.2020
}
